uploading support and privacy docs

+++ EpochTimeConverter/EpochTimeConverter/Model/UNIXconvertion.swift
@@ -11,11 +11,11 @@ class UNIXconvertion : ObservableObject {
        
     // converts the date-time variable into epoch/UNIX time stamp and returns it as a string.
     // type casts the Double to a string so there is no decimal. This creates an output without milliseconds. 
-    func convertToEpoch(date: Date) -> Double{
+    func convertToEpoch(date: Date) -> Int{
         
         let unixTime = date.timeIntervalSince1970.rounded()
         
-        return unixTime
+        return Int(unixTime)
         
     }
